fix(postgres): map ZodNumber to INTEGER instead of NUMERIC
postgres.js returns NUMERIC values as strings, causing Zod z.number()
validation to fail on read-back. INTEGER values are returned as JS
numbers. All numeric fields in the project are integers.

+++ src/adapters/postgres/read-model.ts
@@ -36,7 +36,7 @@ function zodToColumnType(zodType: z.ZodTypeAny): string {
     return "TEXT";
   }
 
-  if (typeName === "ZodNumber") return "NUMERIC";
+  if (typeName === "ZodNumber") return "INTEGER";
   if (typeName === "ZodBoolean") return "BOOLEAN";
   if (typeName === "ZodArray" || typeName === "ZodObject") return "JSONB";
